+++ store/sqlstore/preference_store.go
@@ -259,34 +259,37 @@ func (s SqlPreferenceStore) DeleteCategoryAndName(category string, name string) 
 	return nil
 }
 
-func (s SqlPreferenceStore) CleanupFlagsBatch(limit int64) (int64, error) {
-	query :=
-		`DELETE FROM
-			Preferences
-		WHERE
-			Category = :Category
-			AND Name IN (
-				SELECT
-					*
-				FROM (
-					SELECT
-						Preferences.Name
-					FROM
-						Preferences
-					LEFT JOIN
-						Posts
-					ON
-						Preferences.Name = Posts.Id
-					WHERE
-						Preferences.Category = :Category
-						AND Posts.Id IS null
-					LIMIT
-						:Limit
-				)
-				AS t
-			)`
+func (s SqlPreferenceStore) PermanentDeleteFlagsBatch(endTime, limit int64) (int64, error) {
+	// Granular policies override global ones
+	const selectQuery = `
+		SELECT Preferences.Name FROM Preferences
+		LEFT JOIN Posts ON Preferences.Name = Posts.Id
+		LEFT JOIN Channels ON Posts.ChannelId = Channels.Id
+		LEFT JOIN Teams ON Channels.TeamId = Teams.Id
+		LEFT JOIN RetentionPoliciesChannels ON Posts.ChannelId = RetentionPoliciesChannels.ChannelId
+		LEFT JOIN RetentionPoliciesTeams ON Channels.TeamId = RetentionPoliciesTeams.TeamId
+		WHERE Category = :Category
+		      AND RetentionPoliciesChannels.ChannelId IS NULL
+		      AND RetentionPoliciesTeams.TeamId IS NULL
+		      AND Posts.CreateAt < :EndTime
+		LIMIT :Limit`
+	var query string
+	if s.DriverName() == model.DATABASE_DRIVER_POSTGRES {
+		query = `
+		DELETE FROM Preferences
+		WHERE Name IN (
+		` + selectQuery + `
+		)`
+	} else {
+		// MySQL does not support the LIMIT clause in a subquery with IN
+		query = `
+		DELETE Preferences FROM Preferences INNER JOIN (
+		` + selectQuery + `
+		) AS A ON Preferences.Name = A.Name`
+	}
 
-	sqlResult, err := s.GetMaster().Exec(query, map[string]interface{}{"Category": model.PREFERENCE_CATEGORY_FLAGGED_POST, "Limit": limit})
+	props := map[string]interface{}{"Category": model.PREFERENCE_CATEGORY_FLAGGED_POST, "EndTime": endTime, "Limit": limit}
+	sqlResult, err := s.GetMaster().Exec(query, props)
 	if err != nil {
 		return int64(0), errors.Wrap(err, "failed to delete Preference")
 	}
@@ -299,34 +302,43 @@ func (s SqlPreferenceStore) CleanupFlagsBatch(limit int64) (int64, error) {
 	return rowsAffected, nil
 }
 
-func (s *SqlPreferenceStore) CleanupFlagsBatchForRetentionPolicies(now int64, limit int64) (int64, error) {
+func (s *SqlPreferenceStore) PermanentDeleteFlagsBatchForRetentionPolicies(now int64, limit int64) (int64, error) {
 	// Channel-specific policies override team-specific policies.
 	// This will delete a preference if its corresponding post was already deleted was by
 	// a retention policy.
-	const sQuery = `
-	DELETE FROM Preferences 
-	WHERE Category = :Category
-	AND Name IN (
+	const selectQuery = `
 		SELECT Preferences.Name FROM Preferences
 		LEFT JOIN Posts ON Preferences.Name = Posts.Id
 		LEFT JOIN Channels ON Posts.ChannelId = Channels.Id
 		LEFT JOIN RetentionPoliciesChannels ON Posts.ChannelId = RetentionPoliciesChannels.ChannelId
 		LEFT JOIN RetentionPoliciesTeams ON Channels.TeamId = RetentionPoliciesTeams.TeamId
-		LEFT JOIN RetentionPolicies ON RetentionPoliciesChannels.PolicyId = RetentionPolicies.Id
-			OR RetentionPoliciesTeams.PolicyId = RetentionPolicies.Id
-		WHERE (
-			RetentionPoliciesChannels.ChannelId IS NOT NULL
-			AND :Now - Posts.CreateAt >= RetentionPolicies.PostDuration * 24 * 60 * 60 * 1000
-		) OR (
-			RetentionPoliciesTeams.TeamId IS NOT NULL AND RetentionPoliciesChannels.ChannelId IS NULL
-			AND :Now - Posts.CreateAt >= RetentionPolicies.PostDuration * 24 * 60 * 60 * 1000
-		) OR (
-			Posts.Id IS NULL
+		LEFT JOIN RetentionPolicies ON
+			RetentionPoliciesChannels.PolicyId = RetentionPolicies.Id
+			OR (
+				RetentionPoliciesChannels.PolicyId IS NULL
+				AND RetentionPoliciesTeams.PolicyId = RetentionPolicies.Id
+			)
+		WHERE Category = :Category AND (
+			:Now - Posts.CreateAt >= RetentionPolicies.PostDuration * 24 * 60 * 60 * 1000
+			OR Posts.Id IS NULL
 		)
-		LIMIT :Limit
-	)`
+		LIMIT :Limit`
+	var query string
+	if s.DriverName() == model.DATABASE_DRIVER_POSTGRES {
+		query = `
+		DELETE FROM Preferences
+		WHERE Name IN (
+		` + selectQuery + `
+		)`
+	} else {
+		// MySQL does not support the LIMIT clause in a subquery with IN
+		query = `
+		DELETE Preferences FROM Preferences INNER JOIN (
+		` + selectQuery + `
+		) AS A ON Preferences.Name = A.Name`
+	}
 	props := map[string]interface{}{"Category": model.PREFERENCE_CATEGORY_FLAGGED_POST, "Now": now, "Limit": limit}
-	result, err := s.GetMaster().Exec(sQuery, props)
+	result, err := s.GetMaster().Exec(query, props)
 	if err != nil {
 		return 0, errors.Wrap(err, "failed to delete Preferences")
 	}
